requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
langgraph==0.0.25
pydantic==2.5.2
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.1.1
langchain-openai==0.0.5
python-dotenv==1.0.0 